Change type of magic field to long

+++ src/main/java/cloud/metaapi/sdk/clients/meta_api/models/MetatraderAccountDto.java
@@ -68,7 +68,7 @@ public class MetatraderAccountDto {
   /**
    * MetaTrader magic to place trades using
    */
-  public int magic;
+  public long magic;
   /**
    * Account deployment state
    */
