+++ src/main/java/cloud/metaapi/sdk/clients/meta_api/models/MetatraderOrder.java
@@ -55,7 +55,7 @@ public class MetatraderOrder {
     /**
      * Order magic number, identifies the EA which created the order
      */
-    public int magic;
+    public long magic;
     /**
      * Time order was created at
      */
